Moves all items into a single inner package
