+++ app/src/main/java/com/uncc/gamejam49er/espionage/items/TrollWatch.java
@@ -1,4 +1,4 @@
-package com.uncc.gamejam49er.espionage;
+package com.uncc.gamejam49er.espionage.items;
 
 /**
  * Created by Megan on 10/24/2015.
